pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "histomoe"
version = "0.1.0"
description = "HistoMoE: A Histology-Guided Mixture-of-Experts Framework for Gene Expression Prediction"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    { name = "HistoMoE Contributors" }
]
keywords = [
    "computational pathology",
    "spatial transcriptomics",
    "mixture of experts",
    "gene expression prediction",
    "multimodal learning",
    "deep learning"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "transformers>=4.30.0",
    "scanpy>=1.9.0",
    "anndata>=0.9.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "Pillow>=9.4.0",
    "h5py>=3.8.0",
    "scipy>=1.10.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.2.0",
    "ipykernel>=6.23.0",
    "jupyter>=1.0.0",
]
notebook = [
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
    "plotly>=5.14.0",
]

[project.scripts]
histomoe-train = "train:main"
histomoe-evaluate = "evaluate:main"

[project.urls]
"Homepage" = "https://github.com/your-org/histomoe"
"Bug Reports" = "https://github.com/your-org/histomoe/issues"
"Source" = "https://github.com/your-org/histomoe"
"Documentation" = "https://histomoe.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["histomoe*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]
